Email is the new username
Feels more straight forward to use only email for username. For my
intentions the account is more or less worthless if there is no way of
contacting the user, so a user should minimally have an email-claim.
However this poses some problems.

- Username is required and used resource owner flow. Setting the email
as the username brings pain upon supporting email change since asp.net
identity is not build to change username. Instead we set it to a guid
and forget about it. We'll use extension-points in identity server to
overcome the issues.